fix initmemberlisttable function

+++ MP1Node.cpp
@@ -105,7 +105,7 @@ int MP1Node::initThisNode(Address *joinaddr) {
 	memberNode->heartbeat = 0;
 	memberNode->pingCounter = TFAIL;
 	memberNode->timeOutCounter = -1;
-    initMemberListTable(memberNode, id, port);
+    initMemberListTable(memberNode, id, (short)port);
 
     return 0;
 }
